restructure the app architecture
