+++ android/local.properties
@@ -1,5 +1,5 @@
-sdk.dir=C:\\Users\\WESLEY\\AppData\\Local\\Android\\Sdk
-flutter.sdk=C:\\Users\\WESLEY\\AppData\\Local\\flutter
+sdk.dir=/Users/yunwen/Library/Android/sdk
+flutter.sdk=/Users/yunwen/fvm/versions/3.24.0
 flutter.buildMode=debug
 flutter.versionName=1.0.0
 flutter.versionCode=1
