+++ lib/authentication/screens/auth_pages.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:shopping_app/screens/login_screen.dart';
-import 'package:shopping_app/screens/register_screen.dart';
+import 'package:shopping_app/authentication/screens/login_screen.dart';
+import 'package:shopping_app/authentication/screens/register_screen.dart';
 
 class AuthPage extends StatefulWidget {
   const AuthPage({super.key});
@@ -9,16 +9,13 @@ class AuthPage extends StatefulWidget {
   State<AuthPage> createState() => _AuthPageState();
 }
 
-class _AuthPageState extends State<AuthPage> with SingleTickerProviderStateMixin{
+class _AuthPageState extends State<AuthPage> with SingleTickerProviderStateMixin {
   late TabController _tabController;
 
   @override
   void initState() {
     super.initState();
-    _tabController = TabController(
-      length: 2,
-      vsync: this
-    );
+    _tabController = TabController(length: 2, vsync: this);
   }
 
   @override
@@ -37,25 +34,18 @@ class _AuthPageState extends State<AuthPage> with SingleTickerProviderStateMixin
           tabs: const [
             Padding(
               padding: EdgeInsets.all(8.0),
-              child: Text(
-                "Log In"
-              ),
+              child: Text("Log In"),
             ),
             Padding(
               padding: EdgeInsets.all(8.0),
-              child: Text(
-                "Sign Up"
-              ),
+              child: Text("Sign Up"),
             ),
           ],
         ),
       ),
       body: TabBarView(
         controller: _tabController,
-        children: const [
-          LoginScreen(),
-          RegisterScreen()
-        ],
+        children: const [LoginScreen(), RegisterScreen()],
       ),
     );
   }
