+++ lib/products/data/models/product.dart
