+++ lib/shared/util/grid_box.dart
@@ -1,17 +1,16 @@
 import 'package:flutter/material.dart';
-import 'package:shopping_app/screens/details_screen.dart';
+import 'package:shopping_app/home/screens/details_screen.dart';
 import "package:cached_network_image/cached_network_image.dart";
 
 class GridBox extends StatelessWidget {
   final String title, imagePath, description;
   final double price, rating;
-  const GridBox({super.key,required this.title, required this.imagePath, required this.price, required this.description, required this.rating});
+  const GridBox({super.key, required this.title, required this.imagePath, required this.price, required this.description, required this.rating});
 
   @override
   Widget build(BuildContext context) {
     return Container(
-      decoration: BoxDecoration(
-          color: Colors.white, borderRadius: BorderRadius.circular(12)),
+      decoration: BoxDecoration(color: Colors.white, borderRadius: BorderRadius.circular(12)),
       child: Column(
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
@@ -33,8 +32,7 @@ class GridBox extends StatelessWidget {
               children: [
                 Text(
                   "\$${price.toStringAsFixed(2)}",
-                  style: const TextStyle(
-                      color: Color.fromARGB(255, 201, 127, 214)),
+                  style: const TextStyle(color: Color.fromARGB(255, 201, 127, 214)),
                 ),
                 IconButton(
                   onPressed: () {
